requirements.txt
fastapi==0.104.1
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
python-slugify==8.0.1
aiofiles==23.2.1
uvicorn==0.24.0
watchdog
Pillow==10.2.0 